# openapi/openapi.yml
components:
  schemas:
...
    Attributes:
      description: "Members of the attributes object (\"attributes\") represent information\
        \ about the resource object in which it's defined.\nThe keys for Attributes\
        \ MUST NOT be:\n    relationships\n    links\n    id\n    type"
      properties: {}
      title: Attributes
      type: object
...
    BaseInfoAttributes:
      description: Attributes for Base URL Info endpoint
      properties:
        api_version:
          description: 'Presently used full version of the OPTIMADE API.

            The version number string MUST NOT be prefixed by, e.g., "v".

            Examples: `1.0.0`, `1.0.0-rc.2`.'
          pattern: ^(0|[1-9]\d*)\.(0|[1-9]\d*)\.(0|[1-9]\d*)(?:-((?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\.(?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\+([0-9a-zA-Z-]+(?:\.[0-9a-zA-Z-]+)*))?$
          title: Api Version
          type: string
        available_api_versions:
          description: A list of dictionaries of available API versions at other base
            URLs
          items:
            $ref: '#/components/schemas/AvailableApiVersion'
          title: Available Api Versions
          type: array
        available_endpoints:
          description: List of available endpoints (i.e., the string to be appended
            to the versioned base URL).
          items:
            type: string
          title: Available Endpoints
          type: array
        entry_types_by_format:
          additionalProperties:
            items:
              type: string
            type: array
          description: Available entry endpoints as a function of output formats.
          title: Entry Types By Format
          type: object
        formats:
          default:
          - json
          description: List of available output formats.
          items:
            type: string
          title: Formats
          type: array
        is_index:
          default: false
          description: If true, this is an index meta-database base URL (see section
            Index Meta-Database). If this member is not provided, the client MUST
            assume this is not an index meta-database base URL (i.e., the default
            is for `is_index` to be `false`).
          title: Is Index
          type: boolean
      required:
      - api_version
      - available_api_versions
      - available_endpoints
      - entry_types_by_format
      title: BaseInfoAttributes
      type: object
    BaseInfoResource:
      description: Resource objects appear in a JSON API document to represent resources.
      properties:
        attributes:
          $ref: '#/components/schemas/BaseInfoAttributes'
        id:
          pattern: ^/$
          title: Id
          type: string
        links:
          allOf:
          - $ref: '#/components/schemas/ResourceLinks'
          description: a links object containing links related to the resource.
          title: Links
        meta:
          allOf:
          - $ref: '#/components/schemas/Meta'
          description: a meta object containing non-standard meta-information about
            a resource that can not be represented as an attribute or relationship.
          title: Meta
        relationships:
          allOf:
          - $ref: '#/components/schemas/Relationships'
          description: '[Relationships object](https://jsonapi.org/format/1.0/#document-resource-object-relationships)

            describing relationships between the resource and other JSON API resources.'
          title: Relationships
        type:
          pattern: ^info$
          title: Type
          type: string
      required:
      - id
      - type
      - attributes
      title: BaseInfoResource
      type: object
...
    EntryInfoResource:
      properties:
        description:
          description: Description of the entry.
          title: Description
          type: string
        formats:
          description: List of output formats available for this type of entry.
          items:
            type: string
          title: Formats
          type: array
        output_fields_by_format:
          additionalProperties:
            items:
              type: string
            type: array
          description: Dictionary of available output fields for this entry type,
            where the keys are the values of the `formats` list and the values are
            the keys of the `properties` dictionary.
          title: Output Fields By Format
          type: object
        properties:
          additionalProperties:
            $ref: '#/components/schemas/EntryInfoProperty'
          description: A dictionary describing queryable properties for this entry
            type, where each key is a property name.
          title: Properties
          type: object
      required:
      - formats
      - description
      - properties
      - output_fields_by_format
      title: EntryInfoResource
      type: object
    EntryInfoResponse:
      description: errors are not allowed
      properties:
        data:
          allOf:
          - $ref: '#/components/schemas/EntryInfoResource'
          description: OPTIMADE information for an entry endpoint.
          title: Data
        errors:
          description: A list of unique errors
          items:
            $ref: '#/components/schemas/Error'
          title: Errors
          type: array
          uniqueItems: true
        included:
          description: A list of unique included resources
          items:
            $ref: '#/components/schemas/Resource'
          title: Included
          type: array
          uniqueItems: true
        jsonapi:
          allOf:
          - $ref: '#/components/schemas/JsonApi'
          description: Information about the JSON API used
          title: Jsonapi
        links:
          allOf:
          - $ref: '#/components/schemas/ToplevelLinks'
          description: Links associated with the primary data or errors
          title: Links
        meta:
          allOf:
          - $ref: '#/components/schemas/ResponseMeta'
          description: A meta object containing non-standard information
          title: Meta
      required:
      - data
      - meta
      title: EntryInfoResponse
      type: object
...
    Error:
      description: An error response
      properties:
        code:
          description: an application-specific error code, expressed as a string value.
          title: Code
          type: string
        detail:
          description: A human-readable explanation specific to this occurrence of
            the problem.
          title: Detail
          type: string
        id:
          description: A unique identifier for this particular occurrence of the problem.
          title: Id
          type: string
        links:
          allOf:
          - $ref: '#/components/schemas/ErrorLinks'
          description: A links object storing about
          title: Links
        meta:
          allOf:
          - $ref: '#/components/schemas/Meta'
          description: a meta object containing non-standard meta-information about
            the error.
          title: Meta
        source:
          allOf:
          - $ref: '#/components/schemas/ErrorSource'
          description: An object containing references to the source of the error
          title: Source
        status:
          description: the HTTP status code applicable to this problem, expressed
            as a string value.
          title: Status
          type: string
        title:
          description: A short, human-readable summary of the problem. It **SHOULD
            NOT** change from occurrence to occurrence of the problem, except for
            purposes of localization.
          title: Title
          type: string
      title: Error
      type: object
    ErrorLinks:
      description: A Links object specific to Error objects
      properties:
        about:
          anyOf:
          - format: uri
            maxLength: 65536
            minLength: 1
            type: string
          - $ref: '#/components/schemas/Link'
          description: A link that leads to further details about this particular
            occurrence of the problem.
          title: About
      title: ErrorLinks
      type: object
    ErrorResponse:
      description: errors MUST be present and data MUST be skipped
      properties:
        data:
          anyOf:
          - $ref: '#/components/schemas/Resource'
          - items:
              $ref: '#/components/schemas/Resource'
            type: array
          description: Outputted Data
          title: Data
          uniqueItems: true
        errors:
          description: A list of OPTIMADE-specific JSON API error objects, where the
            field detail MUST be present.
          items:
            $ref: '#/components/schemas/OptimadeError'
          title: Errors
          type: array
          uniqueItems: true
        included:
          description: A list of unique included resources
          items:
            $ref: '#/components/schemas/Resource'
          title: Included
          type: array
          uniqueItems: true
        jsonapi:
          allOf:
          - $ref: '#/components/schemas/JsonApi'
          description: Information about the JSON API used
          title: Jsonapi
        links:
          allOf:
          - $ref: '#/components/schemas/ToplevelLinks'
          description: Links associated with the primary data or errors
          title: Links
        meta:
          allOf:
          - $ref: '#/components/schemas/ResponseMeta'
          description: A meta object containing non-standard information.
          title: Meta
      required:
      - meta
      - errors
      title: ErrorResponse
      type: object
    ErrorSource:
      description: an object containing references to the source of the error
      properties:
        parameter:
          description: a string indicating which URI query parameter caused the error.
          title: Parameter
          type: string
        pointer:
          description: a JSON Pointer [RFC6901] to the associated entity in the request
            document [e.g. "/data" for a primary data object, or "/data/attributes/title"
            for a specific attribute].
          title: Pointer
          type: string
      title: ErrorSource
      type: object
...
    Implementation:
      description: Information on the server implementation
      properties:
        homepage:
          anyOf:
          - format: uri
            maxLength: 65536
            minLength: 1
            type: string
          - $ref: '#/components/schemas/Link'
          description: A [JSON API links object](http://jsonapi.org/format/1.0/#document-links)
            pointing to the homepage of the implementation.
          title: Homepage
        issue_tracker:
          anyOf:
          - format: uri
            maxLength: 65536
            minLength: 1
            type: string
          - $ref: '#/components/schemas/Link'
          description: A [JSON API links object](http://jsonapi.org/format/1.0/#document-links)
            pointing to the implementation's issue tracker.
          title: Issue Tracker
        maintainer:
          allOf:
          - $ref: '#/components/schemas/ImplementationMaintainer'
          description: A dictionary providing details about the maintainer of the
            implementation.
          title: Maintainer
        name:
          description: name of the implementation
          title: Name
          type: string
        source_url:
          anyOf:
          - format: uri
            maxLength: 65536
            minLength: 1
            type: string
          - $ref: '#/components/schemas/Link'
          description: A [JSON API links object](http://jsonapi.org/format/1.0/#document-links)
            pointing to the implementation source, either downloadable archive or
            version control system.
          title: Source Url
        version:
          description: version string of the current implementation
          title: Version
          type: string
      title: Implementation
      type: object
...
    InfoResponse:
      description: errors are not allowed
      properties:
        data:
          allOf:
          - $ref: '#/components/schemas/BaseInfoResource'
          description: The implementations /info data.
          title: Data
        errors:
          description: A list of unique errors
          items:
            $ref: '#/components/schemas/Error'
          title: Errors
          type: array
          uniqueItems: true
        included:
          description: A list of unique included resources
          items:
            $ref: '#/components/schemas/Resource'
          title: Included
          type: array
          uniqueItems: true
        jsonapi:
          allOf:
          - $ref: '#/components/schemas/JsonApi'
          description: Information about the JSON API used
          title: Jsonapi
        links:
          allOf:
          - $ref: '#/components/schemas/ToplevelLinks'
          description: Links associated with the primary data or errors
          title: Links
        meta:
          allOf:
          - $ref: '#/components/schemas/ResponseMeta'
          description: A meta object containing non-standard information
          title: Meta
      required:
      - data
      - meta
      title: InfoResponse
      type: object
    JsonApi:
      description: An object describing the server's implementation
      properties:
        meta:
          allOf:
          - $ref: '#/components/schemas/Meta'
          description: Non-standard meta information
          title: Meta
        version:
          default: '1.0'
          description: Version of the json API used
          title: Version
          type: string
      title: JsonApi
      type: object
    Link:
      description: 'A link **MUST** be represented as either: a string containing
        the link''s URL or a link object.'
      properties:
        href:
          description: "a string containing the link\u2019s URL."
          format: uri
          maxLength: 65536
          minLength: 1
          title: Href
          type: string
        meta:
          allOf:
          - $ref: '#/components/schemas/Meta'
          description: a meta object containing non-standard meta-information about
            the link.
          title: Meta
      required:
      - href
      title: Link
      type: object
...
    Meta:
      description: Non-standard meta-information that can not be represented as an
        attribute or relationship.
      properties: {}
      title: Meta
      type: object
...
    OptimadeError:
      description: detail MUST be present
      properties:
        code:
          description: an application-specific error code, expressed as a string value.
          title: Code
          type: string
        detail:
          description: A human-readable explanation specific to this occurrence of
            the problem.
          title: Detail
          type: string
        id:
          description: A unique identifier for this particular occurrence of the problem.
          title: Id
          type: string
        links:
          allOf:
          - $ref: '#/components/schemas/ErrorLinks'
          description: A links object storing about
          title: Links
        meta:
          allOf:
          - $ref: '#/components/schemas/Meta'
          description: a meta object containing non-standard meta-information about
            the error.
          title: Meta
        source:
          allOf:
          - $ref: '#/components/schemas/ErrorSource'
          description: An object containing references to the source of the error
          title: Source
        status:
          description: the HTTP status code applicable to this problem, expressed
            as a string value.
          title: Status
          type: string
        title:
          description: A short, human-readable summary of the problem. It **SHOULD
            NOT** change from occurrence to occurrence of the problem, except for
            purposes of localization.
          title: Title
          type: string
      required:
      - detail
      title: OptimadeError
      type: object
...
    Provider:
      description: Information on the database provider of the implementation.
      properties:
        description:
          description: a longer description of the database provider
          title: Description
          type: string
        homepage:
          anyOf:
          - format: uri
            maxLength: 65536
            minLength: 1
            type: string
          - $ref: '#/components/schemas/Link'
          description: a [JSON API links object](http://jsonapi.org/format/1.0#document-links)
            pointing to homepage of the database provider, either directly as a string,
            or as a link object.
          title: Homepage
        name:
          description: a short name for the database provider
          title: Name
          type: string
        prefix:
          description: database-provider-specific prefix as found in section Database-Provider-Specific
            Namespace Prefixes.
          pattern: ^[a-z]([a-z]|[0-9]|_)*$
          title: Prefix
          type: string
      required:
      - name
      - description
      - prefix
      title: Provider
      type: object
...
    Relationships:
      description: "Members of the relationships object (\"relationships\") represent\
        \ references from the resource object in which it's defined to other resource\
        \ objects.\nKeys MUST NOT be:\n    type\n    id"
      properties: {}
      title: Relationships
      type: object
    Resource:
      description: Resource objects appear in a JSON API document to represent resources.
      properties:
        attributes:
          allOf:
          - $ref: '#/components/schemas/Attributes'
          description: "an attributes object representing some of the resource\u2019\
            s data."
          title: Attributes
        id:
          description: Resource ID
          title: Id
          type: string
        links:
          allOf:
          - $ref: '#/components/schemas/ResourceLinks'
          description: a links object containing links related to the resource.
          title: Links
        meta:
          allOf:
          - $ref: '#/components/schemas/Meta'
          description: a meta object containing non-standard meta-information about
            a resource that can not be represented as an attribute or relationship.
          title: Meta
        relationships:
          allOf:
          - $ref: '#/components/schemas/Relationships'
          description: '[Relationships object](https://jsonapi.org/format/1.0/#document-resource-object-relationships)

            describing relationships between the resource and other JSON API resources.'
          title: Relationships
        type:
          description: Resource type
          title: Type
          type: string
      required:
      - id
      - type
      title: Resource
      type: object
    ResourceLinks:
      description: A Resource Links object
      properties:
        self:
          anyOf:
          - format: uri
            maxLength: 65536
            minLength: 1
            type: string
          - $ref: '#/components/schemas/Link'
          description: A link that identifies the resource represented by the resource
            object.
          title: Self
      title: ResourceLinks
      type: object
    ResponseMeta:
      description: 'A [JSON API meta member](https://jsonapi.org/format/1.0#document-meta)

        that contains JSON API meta objects of non-standard

        meta-information.


        OPTIONAL additional information global to the query that is not

        specified in this document, MUST start with a

        database-provider-specific prefix.'
      properties:
        api_version:
          description: 'Presently used full version of the OPTIMADE API.

            The version number string MUST NOT be prefixed by, e.g., "v".

            Examples: `1.0.0`, `1.0.0-rc.2`.'
          pattern: ^(0|[1-9]\d*)\.(0|[1-9]\d*)\.(0|[1-9]\d*)(?:-((?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\.(?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\+([0-9a-zA-Z-]+(?:\.[0-9a-zA-Z-]+)*))?$
          title: Api Version
          type: string
        data_available:
          description: An integer containing the total number of data resource objects
            available in the database for the endpoint.
          title: Data Available
          type: integer
        data_returned:
          description: An integer containing the total number of data resource objects
            returned for the current `filter` query, independent of pagination.
          minimum: 0.0
          title: Data Returned
          type: integer
        implementation:
          allOf:
          - $ref: '#/components/schemas/Implementation'
          description: a dictionary describing the server implementation
          title: Implementation
        last_id:
          description: a string containing the last ID returned
          title: Last Id
          type: string
        more_data_available:
          description: '`false` if the response contains all data for the request
            (e.g., a request issued to a single entry endpoint, or a `filter` query
            at the last page of a paginated response) and `true` if the response is
            incomplete in the sense that multiple objects match the request, and not
            all of them have been included in the response (e.g., a query with multiple
            pages that is not at the last page).'
          title: More Data Available
          type: boolean
        provider:
          allOf:
          - $ref: '#/components/schemas/Provider'
          description: information on the database provider of the implementation.
          title: Provider
        query:
          allOf:
          - $ref: '#/components/schemas/ResponseMetaQuery'
          description: Information on the Query that was requested
          title: Query
        response_message:
          description: response string from the server
          title: Response Message
          type: string
        schema:
          anyOf:
          - format: uri
            maxLength: 65536
            minLength: 1
            type: string
          - $ref: '#/components/schemas/Link'
          description: 'A [JSON API links object](http://jsonapi.org/format/1.0/#document-links)
            that points to a schema for the response.

            If it is a string, or a dictionary containing no `meta` field, the provided
            URL MUST point at an [OpenAPI](https://swagger.io/specification/) schema.

            It is possible that future versions of this specification allows for alternative
            schema types.

            Hence, if the `meta` field of the JSON API links object is provided and
            contains a field `schema_type` that is not equal to the string `OpenAPI`
            the client MUST not handle failures to parse the schema or to validate
            the response against the schema as errors.'
          title: Schema
        time_stamp:
          description: A timestamp containing the date and time at which the query
            was executed.
          format: date-time
          title: Time Stamp
          type: string
        warnings:
          description: 'A list of warning resource objects representing non-critical
            errors or warnings.

            A warning resource object is defined similarly to a [JSON API error object](http://jsonapi.org/format/1.0/#error-objects),
            but MUST also include the field `type`, which MUST have the value `"warning"`.

            The field `detail` MUST be present and SHOULD contain a non-critical message,
            e.g., reporting unrecognized search attributes or deprecated features.

            The field `status`, representing a HTTP response status code, MUST NOT
            be present for a warning resource object.

            This is an exclusive field for error resource objects.'
          items:
            $ref: '#/components/schemas/Warnings'
          title: Warnings
          type: array
          uniqueItems: true
      required:
      - query
      - api_version
      - more_data_available
      title: ResponseMeta
      type: object
    ResponseMetaQuery:
      description: Information on the query that was requested.
      properties:
        representation:
          description: 'A string with the part of the URL following the versioned
            or unversioned base URL that serves the API.

            Query parameters that have not been used in processing the request MAY
            be omitted.

            In particular, if no query parameters have been involved in processing
            the request, the query part of the URL MAY be excluded.

            Example: `/structures?filter=nelements=2`'
          title: Representation
          type: string
      required:
      - representation
      title: ResponseMetaQuery
      type: object
...
    ToplevelLinks:
      description: A set of Links objects, possibly including pagination
      properties:
        first:
          anyOf:
          - format: uri
            maxLength: 65536
            minLength: 1
            type: string
          - $ref: '#/components/schemas/Link'
          description: The first page of data
          title: First
        last:
          anyOf:
          - format: uri
            maxLength: 65536
            minLength: 1
            type: string
          - $ref: '#/components/schemas/Link'
          description: The last page of data
          title: Last
        next:
          anyOf:
          - format: uri
            maxLength: 65536
            minLength: 1
            type: string
          - $ref: '#/components/schemas/Link'
          description: The next page of data
          title: Next
        prev:
          anyOf:
          - format: uri
            maxLength: 65536
            minLength: 1
            type: string
          - $ref: '#/components/schemas/Link'
          description: The previous page of data
          title: Prev
        related:
          anyOf:
          - format: uri
            maxLength: 65536
            minLength: 1
            type: string
          - $ref: '#/components/schemas/Link'
          description: A related resource link
          title: Related
        self:
          anyOf:
          - format: uri
            maxLength: 65536
            minLength: 1
            type: string
          - $ref: '#/components/schemas/Link'
          description: A link to itself
          title: Self
      title: ToplevelLinks
      type: object
...
  securitySchemes:
    OAuth2AuthorizationCodeBearer:
      flows:
        authorizationCode:
          authorizationUrl: https://staging.the-marketplace.eu/oauth/oauth2/auth
          scopes: {}
          tokenUrl: https://staging.the-marketplace.eu/oauth/oauth2/token
      type: oauth2
info:
  description: A gateway server to query multiple OPTIMADE databases.
  title: OPTIMADE Gateway
  version: 0.2.2
  x-application-id: ''
  x-application-name: OPTIMADE Gateway
  x-external-hostname: http://localhost:8000
  x-products:
  - name: OPTIMADE Gateway
    productId: 22e6e756-7179-47c9-8f37-1ee06c7b38ae
openapi: 3.0.2
paths:
...
  /info:
    get:
      description: '`GET /info`


        An introspective endpoint for the gateway service.'
      operationId: get_info_info_get
      responses:
        '200':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/InfoResponse'
          description: Successful Response
        '400':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '422':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '553':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Version Not Supported
      security:
      - OAuth2AuthorizationCodeBearer: []
      summary: Get Info
      tags:
      - Info
  /info/{entry}:
    get:
      description: '`GET /info/{entry}`


        Get information about the gateway service''s entry-listing endpoints.'
      operationId: get_entry_info_info__entry__get
      parameters:
      - in: path
        name: entry
        required: true
        schema:
          title: Entry
          type: string
      responses:
        '200':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/EntryInfoResponse'
          description: Successful Response
        '400':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '422':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '553':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Version Not Supported
      security:
      - OAuth2AuthorizationCodeBearer: []
      summary: Get Entry Info
      tags:
      - Info
...
servers:
- url: http://localhost:8000